Move AllowedMentions class to rest module
