+++ core/src/main/java/discord4j/core/spec/MessageCreateSpec.java
@@ -19,7 +19,7 @@ package discord4j.core.spec;
 import discord4j.core.object.Embed;
 import discord4j.core.object.entity.Attachment;
 import discord4j.core.object.entity.Message;
-import discord4j.core.object.entity.channel.AllowedMentions;
+import discord4j.rest.util.AllowedMentions;
 import discord4j.core.object.entity.channel.MessageChannel;
 import discord4j.discordjson.json.AllowedMentionsData;
 import discord4j.discordjson.json.EmbedData;
