+++ core/src/test/java/discord4j/core/ExampleAllowedMentions.java
@@ -3,7 +3,7 @@ package discord4j.core;
 import discord4j.core.event.domain.message.MessageCreateEvent;
 import discord4j.core.object.entity.Message;
 import discord4j.core.object.entity.User;
-import discord4j.core.object.entity.channel.AllowedMentions;
+import discord4j.rest.util.AllowedMentions;
 import reactor.core.publisher.Mono;
 
 public class ExampleAllowedMentions {
